backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
Pillow==10.1.0
langchain==0.0.335
langgraph==0.0.62
openai==1.3.5
twilio==8.10.0
PyPDF2==3.0.1
aiohttp==3.9.1
anthropic==0.34.2 
pytesseract==0.3.10
aiohttp==3.9.1 
